requirements.txt
Django==5.1.4
Pillow==11.0.0
python-dotenv==1.0.1
dj-inmemorystorage==2.1.0
django-environ==0.11.2
psycopg2
